base-files: simplify /etc/skel/.bashrc
There was a lot in here that didn't need to be in here. Mostly I did
this to kill the setting of PS1 which happens in /etc/profile. The
request was to see the username you are logged in as and /etc/profile
does that and this file overrode that.