linux-yocto: fix RTC devices
The battery backed RTC we have on board is a STM M41T81. The RTC in the
SoC is not battery backed so we need to set the system to load its
date/time info from the battery backed RTC. Previously the driver for
the M41T81 wasn't even being built but instead a different RTC was being
built which is wrong.